# doc/swagger.yaml
openapi: 3.0.0
info:
  description: SSO HTTP API文档.
  version: 1.0.0
  title: SSO Http API
servers:
  - description: 本地
    url: 'http://localhost/api'
  - description: 测试网
    url: 'http://sso-alpha.dbjtech.com/api'
paths:
...
  /role/permission:
    get:
      tags:
        - 权限组与权限的关系
      summary: 查看权限组拥有的权限
      responses:
        '200':
          description: 请求成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    description: 权限ID
                  path:
                    type: string
                    description: 权限的HTTP请求路径
                  method:
                    type: string
                    description: 权限的HTTP请求方法
      security:
        - cookieAuth: []
    put:
      tags:
        - 权限组与权限的关系
      summary: 更新权限组与权限
      description: 清除之前的关系，以上传的为最终结果
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                role_id:
                  type: string
                  description: 权限组ID
                permission_id_list:
                  type: string
                  description: 权限ID集合
      responses:
        '200':
          description: 请求成功
      security:
        - cookieAuth: []
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: EGG_SESS